feat(users): add ListUsers RPC
